+++ Apps/APN-UIKit/APN UIKit/Base.lproj/Main.storyboard
@@ -181,7 +181,7 @@
                                             <stackView opaque="NO" contentMode="scaleToFill" axis="vertical" alignment="center" spacing="16" translatesAutoresizingMaskIntoConstraints="NO" id="AIH-WX-Xne">
                                                 <rect key="frame" x="0.0" y="160.66666666666666" width="374" height="496.33333333333337"/>
                                                 <subviews>
-                                                    <view contentMode="scaleToFill" translatesAutoresizingMaskIntoConstraints="NO" id="czF-xs-EEf" userLabel="InlineView 1" customClass="InAppMessageView" customModule="CioMessagingInApp">
+                                                    <view contentMode="scaleToFill" translatesAutoresizingMaskIntoConstraints="NO" id="czF-xs-EEf" userLabel="InlineView 1" customClass="InlineMessageUIView" customModule="CioMessagingInApp">
                                                         <rect key="frame" x="0.0" y="0.0" width="374" height="50"/>
                                                         <color key="backgroundColor" systemColor="systemBackgroundColor"/>
                                                         <constraints>
@@ -378,7 +378,7 @@
                             <view contentMode="scaleToFill" translatesAutoresizingMaskIntoConstraints="NO" id="Mh2-zI-sF9">
                                 <rect key="frame" x="20" y="64" width="374" height="400"/>
                                 <subviews>
-                                    <view contentMode="scaleToFill" translatesAutoresizingMaskIntoConstraints="NO" id="kvM-jQ-73n" customClass="InAppMessageView" customModule="CioMessagingInApp">
+                                    <view contentMode="scaleToFill" translatesAutoresizingMaskIntoConstraints="NO" id="kvM-jQ-73n" customClass="InlineMessageUIView" customModule="CioMessagingInApp">
                                         <rect key="frame" x="0.0" y="0.0" width="374" height="20"/>
                                         <constraints>
                                             <constraint firstAttribute="height" constant="20" id="g15-GG-CD3"/>
