+++ Sources/MessagingInApp/autogenerated/AutoMockable.generated.swift
@@ -497,13 +497,13 @@ public class InAppEventListenerMock: InAppEventListener, Mock {
 }
 
 /**
- Class to easily create a mocked version of the `InAppMessageViewActionDelegate` class.
+ Class to easily create a mocked version of the `InlineMessageUIViewDelegate` class.
  This class is equipped with functions and properties ready for you to mock!
 
  Note: This file is automatically generated. This means the mocks should always be up-to-date and has a consistent API.
  See the SDK documentation to learn the basics behind using the mock classes in the SDK.
  */
-public class InAppMessageViewActionDelegateMock: InAppMessageViewActionDelegate, Mock {
+public class InlineMessageUIViewDelegateMock: InlineMessageUIViewDelegate, Mock {
     /// If *any* interactions done on mock. `true` if any method or property getter/setter called.
     public var mockCalled: Bool = false //
